refactor: Rename client and extract service class
