+++ USca/DbManager/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="DbManager.App"
+﻿<Application x:Class="USca_DbManager.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:DbManager"
+             xmlns:local="clr-namespace:USca_DbManager"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
