+++ USca/DbManager/MainWindow.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="DbManager.MainWindow"
+﻿<Window x:Class="USca_DbManager.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:DbManager"
+        xmlns:local="clr-namespace:USca_DbManager"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
 	<StackPanel Margin="100 0 100 0" VerticalAlignment="Center">
