+++ USca/DbManager/MainWindow.xaml.cs
@@ -1,4 +1,5 @@
-﻿using RestSharp;
+﻿using USca_DbManager.User;
+using RestSharp;
 using System;
 using System.Diagnostics;
 using System.Net.Http;
@@ -7,7 +8,7 @@ using System.Threading;
 using System.Windows;
 using System.Windows.Input;
 
-namespace DbManager
+namespace USca_DbManager
 {
 	public partial class MainWindow : Window
 	{
@@ -25,43 +26,25 @@ namespace DbManager
 
 		private async void DoLogin(string username, string password)
 		{
-			var dto = new
-			{
-				username,
-				password
-			};
-
-			var URL = "http://localhost:5274/api";
-			using var cli = new RestClient(new RestClientOptions(URL));
-
-			var req = new RestRequest("User", Method.Put);
-			req.AddBody(dto);
-			Console.WriteLine(dto);
-
 			btnLogin.IsEnabled = false;
 			Mouse.OverrideCursor = Cursors.Wait;
-			RestResponse response = await cli.ExecuteAsync(req);
-			Mouse.OverrideCursor = null;
-			btnLogin.IsEnabled = true;
 
-			Console.WriteLine(response.Content);
-
-			if (response.StatusCode == System.Net.HttpStatusCode.NotFound)
+			try
+			{
+				var res = await UserService.Login(username, password);
+				MessageBox.Show($"Logged in as {res.username}", "Alert.", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+			}
+			catch (BadCredentialsException)
 			{
 				MessageBox.Show("Username or Password invalid", "Could not log in.", MessageBoxButton.OK, MessageBoxImage.Error);
 				tbPassword.Clear();
 				tbPassword.Focus();
-			} 
-			else if (response.StatusCode == System.Net.HttpStatusCode.OK)
+			}
+			finally
 			{
-				LoginResultDTO result = JsonSerializer.Deserialize<LoginResultDTO>(response.Content!)!;
-				MessageBox.Show($"Logged in as {result.username}", "Alert.", MessageBoxButton.OK, MessageBoxImage.Exclamation);
+				Mouse.OverrideCursor = null;
+				btnLogin.IsEnabled = true;
 			}
 		}
-	}
-
-	class LoginResultDTO
-	{
-		public string username { get; set; }
 	}
 }
