+++ USca/DbManager/User/BadCredentialsException.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace USca_DbManager.User
+{
+    class BadCredentialsException : Exception
+    {
+    }
+}
