+++ USca/DbManager/User/LoginResultDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace USca_DbManager.User
+{
+	class LoginResultDTO
+	{
+		public string username { get; set; }
+	}
+}
