+++ USca/USca.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.5.33516.290
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "USca-Server", "USca-Server\USca-Server.csproj", "{51E69E79-B720-4410-A83C-1E5655400965}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DbManager", "DbManager\DbManager.csproj", "{3AC74432-DFEA-494A-B929-CA381885980F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "USca_DbManager", "DbManager\USca_DbManager.csproj", "{3AC74432-DFEA-494A-B929-CA381885980F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
